requirements.txt
Flask==1.1.1
torch==1.3.1
torchvision==0.4.2
numpy==1.17.3
Pillow==6.2.1
opencv-python-headless
